requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn[standard]>=0.15.0

# Database
motor>=3.0.0
pinecone-client>=2.0.0

# API Integration
aiohttp>=3.8.0
python-dotenv>=0.19.0

# Data Processing
pydantic>=1.8.0
python-dateutil>=2.8.2

# Async Support
asyncio>=3.4.3

# Utilities
PyYAML>=6.0
python-multipart>=0.0.5

# Testing
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0

# Development
black>=21.9b0
isort>=5.9.3
flake8>=3.9.0